requirements.txt
openai>=1.40.0
python-telegram-bot>=21.0
supabase>=2.5.0
httpx>=0.27.0
pypdf>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
